README.md
Analyze the number of opioid deaths in Allegheny County within the dataset. The objective of this analysis is to determine which of the following has a stronger correlation to the number of deaths, age, gender, race, and drug combinations. The research will be used to devise/determine a method to help reduce the number of opioid related deaths.